refactor: extract package for model
