+++ src/main/java/vendingmachine/Application.java
@@ -3,6 +3,9 @@ package vendingmachine;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Arrays;
 import java.util.List;
+import vendingmachine.model.Coin;
+import vendingmachine.model.Coins;
+import vendingmachine.model.Item;
 import vendingmachine.reader.ExchangeAmountReader;
 import vendingmachine.reader.InputMoneyReader;
 import vendingmachine.reader.ItemListReader;
