+++ src/main/java/vendingmachine/model/Coins.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 import java.util.HashMap;
 import java.util.Map;
