+++ src/main/java/vendingmachine/model/Item.java
@@ -1,4 +1,4 @@
-package vendingmachine;
+package vendingmachine.model;
 
 public class Item {
 	private final String name;
